jacobee: Allow compilation of code with cuda
There is a bug in Makefile when specifying LD flags.
Fixing the same in this patch. It also addresses CUDA 9
and Volta compatibility.

Further, removing support for SM20 gencode from the makefile.
This is because with the current combination there is no CUDA
version with which SM20 works.

If anyone has any combination where SM20 works, following line
should be added manually to $(GENCODE_FLAGS) :

GENCODE_SM20 := -gencode arch=compute_20,code=sm_20

Signed-off-by: Devesh Sharma <devesh.sharma@broadcom.com>